web/Letterwall/requirements.txt
Flask==3.0.2
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
Werkzeug==3.0.1
gunicorn==21.2.0
PyJWT==2.8.0 